@@ -11,45 +11,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^19.2.7",
-    "@angular/compiler": "^19.2.7",
-    "@angular/core": "^19.2.7",
-    "@angular/fire": "^17.1.0",
-    "@angular/forms": "^19.2.7",
-    "@angular/platform-browser": "^19.2.7",
-    "@angular/platform-browser-dynamic": "^19.2.7",
-    "@angular/router": "^19.2.7",
+    "@angular/common": "^20.3.15",
+    "@angular/compiler": "^20.3.15",
+    "@angular/core": "^20.3.15",
+    "@angular/fire": "^20.0.1",
+    "@angular/forms": "^20.3.15",
+    "@angular/platform-browser": "^20.3.15",
+    "@angular/platform-browser-dynamic": "^20.3.15",
+    "@angular/router": "^20.3.15",
     "@ng-bootstrap/ng-bootstrap": "^18.0.0",
     "@popperjs/core": "^2.11.8",
     "bootstrap": "^5.3.8",
     "bootstrap-icons": "^1.13.1",
-    "chart.js": "^4.5.0",
-    "firebase": "^10.14.1",
-    "fraction.js": "^5.2.1",
-    "mathjs": "^14.8.2",
+    "chart.js": "^4.5.1",
+    "mathjs": "^15.1.0",
     "rxjs": "~7.8.2",
     "tslib": "^2.8.1",
-    "zone.js": "~0.15.1"
+    "zone.js": "~0.16.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.2.8",
-    "@angular/cli": "^19.2.8",
-    "@angular/compiler-cli": "^19.2.7",
-    "@angular/localize": "^19.2.7",
-    "@stylistic/eslint-plugin": "^5.2.2",
-    "@types/jasmine": "~5.1.7",
-    "@typescript-eslint/eslint-plugin": "^8.44.1",
-    "@typescript-eslint/parser": "^8.44.1",
-    "angular-eslint": "19.3.0",
-    "eslint": "^9.36.0",
-    "jasmine-core": "~5.12.0",
+    "@angular/build": "^20.3.13",
+    "@angular/cli": "^20.3.13",
+    "@angular/compiler-cli": "^20.3.15",
+    "@angular/localize": "^20.3.15",
+    "@stylistic/eslint-plugin": "^5.6.1",
+    "@types/jasmine": "~5.1.13",
+    "@typescript-eslint/eslint-plugin": "^8.48.1",
+    "@typescript-eslint/parser": "^8.48.1",
+    "angular-eslint": "^20.0.0",
+    "eslint": "^9.39.1",
+    "jasmine-core": "~5.13.0",
     "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "npm-check-updates": "^17.1.18",
-    "typescript": "~5.8.3",
-    "typescript-eslint": "8.27.0"
+    "npm-check-updates": "^19.1.2",
+    "typescript": "~5.9.3",
+    "typescript-eslint": "8.48.1"
   }
 }
